break out snippets test file and add navigation test
